Extract shared RenderSegments component for message text
The spoiler/URL/styled-text/plain-text segment branch was duplicated
across three render paths (MessageList's standard and compact layouts,
and LinkedText) — which is exactly how the compact-mode spoiler leak
slipped through: a new segment kind only got handled in two of them.

Move that branch into one RenderSegments.vue. It takes a precomputed
RenderSegment[] plus an optional selfColor and renders the lot, spoiler
branch first. MessageList's two layouts and LinkedText now all funnel
through it, so a future segment kind can only be missed in one place.
Segment *production* still differs per call site (LinkedText splits
plain text with no nick set; MessageList splits with the buffer's nick
context) — only the rendering is shared.
